Refactor the way we structure config file modules
We no longer symlink the modules we want locally into the conf.d
folder. Instead we actually commit a config file including the modules
we want as hosts/<hostname>/config, and then symlink that into config
in the root of the repo (~/.config/sway/config). That way we actually
commit the full configuration state used on the machines.

For now we have only done this for "usbee" but the idea is that we
will do this work for the other hosts as I update those to the latest
version of this repo (because I'm too lazy to run around to all the
relevant machines)

TODO: Maybe make some of the smaller modules inside hosts/usbee inline
into hosts/usbee/config. E.g. hosts/usbee/01-env.conf