50-base: Make a particular unity popup always be floating
Actually this isn't good enough on its own... We need it to be focused
by for one (which only seems to realy work if the original
unity window itself is in floating mode at the time). Another thing
would be to have it's placement be where it actually should be, but
might not be possible because xwayland. Maybe some sort of hack that
places it by the mouse?